requirements.txt
python-qbittorrent
python-dotenv
